requirements.txt
google-generativeai>=0.3.0
mutagen>=1.47.0
python-dotenv>=1.0.0
